Upgrade to stylelint v14.x
Also upgrade downstream dependencies:

* stylelint-config-recommended-scss to 5.x

stylelint-scss is a dependency of stylelint-config-recommended-scss, so
we do not need to include it explicitly.

The `syntax` option is no longer provided by stylelint >= 14.0.0. The
stylelint-config-recommended-scss package provides the same function by
using the `extends` config.

See https://github.com/stylelint/stylelint/blob/HEAD/docs/migration-guide/to-14.md#syntax-option-and-automatic-inferral-of-syntax